workerpool: refactor results to avoid unnecessary copy
Background: To avoid leaking the internal implementation of the task
struct, WorkPool.Drain() return a slice of Task interface.

Before this patch, the WorkPool struct would maintain a slice of concret
task struct pointers to be used as return value for the Drain() method.
Since Drain() returns a slice of Task interface, a conversion was
required.

This patch make it so the WorkPool results is a slice of Task interface
instead. That way it can be returned as-is by Drain().

Signed-off-by: Alexandre Perrin <alex@kaworu.ch>