add context to the task function signature
There are use-cases where the caller does not care about the results of
tasks processing and thus never calls the Drain method. This is
typically the case for tasks that are never expected to return unless a
context is cancelled. To accommodate for this use-case, add a context as
a parameter to the task function.

As tasks are now allowed to run forever unless the context is cancelled,
ensure that calling the Close method cancels the tasks context to
properly tear down all workers. It is the responsibility of the library
user to ensure submitted tasks respect the context.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>