Move tests into a separate package
Moving the test code out of the workerpool package allows to write tests
from the perspective of a real user of the package. We cannot possibly
fiddle around with internals and keep tests focussed on the exposed API.

Currently, there's still one case where we want to check internal state:
the (*WorkerPool).tasks capacity check. For that, export a
(*WorkerPool).TasksCap method for testing only.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>