Avoid crash when running task with (*task).run == nil
The following program currently leads to a crash due to nil pointer
dereference:

	wp := workerpool.New(runtime.NumCPU())
	defer wp.Close()
	if err := wp.Submit("foobar", nil); err != nil {
		panic(err)
	}
	tasks, err := wp.Drain()

Fix this by checking (*task).run before running it. This way we keep the
semantics of returning a result for every submitted task. The
alternative solution would be to not enqueue the task on
(*WorkerPool).Submit("foobar", nil) and return either nil or an error.
Both would lead to an imbalance between bumber of submitted tasks and
returned results.

Signed-off-by: Tobias Klauser <tobias@cilium.io>