add NewWithContext
Before this patch, forwarding context cancellation to the workerpool
tasks was difficult and required to spawn a goroutine to call wp.Close
when the parent context was done.

This patch introduce NewWithContext allowing callers to provide a parent
context from which is derived the tasks context, removing the need to
manually forward context cancellation.

Signed-off-by: Alexandre Perrin <alex@isovalent.com>